Expansions done. IFS discarted. Norminette friendly.
