+++ src/expansions/double_quote_utils.c
@@ -6,7 +6,7 @@
 /*   By: jgirbau- <jgirbau-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 16:24:30 by jgirbau-          #+#    #+#             */
-/*   Updated: 2025/10/21 18:08:45 by jgirbau-         ###   ########.fr       */
+/*   Updated: 2025/10/21 19:28:56 by jgirbau-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
