+++ src/expansions/expansions.c
@@ -6,7 +6,7 @@
 /*   By: jgirbau- <jgirbau-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 18:15:06 by jgirbau-          #+#    #+#             */
-/*   Updated: 2025/10/21 12:19:31 by jgirbau-         ###   ########.fr       */
+/*   Updated: 2025/10/21 19:34:32 by jgirbau-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,65 +92,3 @@ char	**expand(char **args, t_env *env)
 	}
 	return (args);
 }
-
-// char	*respect_quotes(char *args)
-// {
-// 	int	i;
-// 	int	j;
-// 	int	k;
-
-// 	if (!args)
-// 		return (NULL);
-// 	i = 0;
-// 	while (args[i])
-// 	{
-// 		if (args[i] == '\'')
-// 		{
-// 			j = find_closure(args, '\'', i);
-// 			if (j)
-// 			{
-// 				k = i + 1;
-// 				while (k < j)
-// 				{
-// 					if (args[k] == '"')
-// 						args[k] = 5;
-// 					k++;
-// 				}
-// 				i = j + 1;
-// 				continue ;
-// 			}
-// 		}
-// 		else if (args[i] == '"')
-// 		{
-// 			j = find_closure(args, '"', i);
-// 			if (j)
-// 			{
-// 				k = i + 1;
-// 				while (k < j)
-// 				{
-// 					if (args[k] == '\'')
-// 						args[k] = 6;
-// 					k++;
-// 				}
-// 				i = j + 1;
-// 				continue ;
-// 			}
-// 		}
-// 		i++;
-// 	}
-// 	return (args);
-// }
-
-// int	find_n(int *n)
-// {
-// 	int	i;
-
-// 	i = 0;
-// 	if (n[i] == 0)
-// 		return ('$');
-// 	else if (n[i] == 1)
-// 		return ('\'');
-// 	else if (n[i] == 2)
-// 		return ('\"');
-// 	return (3);
-// }
