+++ src/expansions/no_quote.c
@@ -6,87 +6,12 @@
 /*   By: jgirbau- <jgirbau-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/09 18:33:32 by jgirbau-          #+#    #+#             */
-/*   Updated: 2025/10/21 17:36:44 by jgirbau-         ###   ########.fr       */
+/*   Updated: 2025/10/21 19:28:04 by jgirbau-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minishell.h"
 #include "../../inc/expansion.h"
-
-char	**case_no_ws(char *dollar)
-{
-	char	**split;
-
-	split = malloc(sizeof(char *) * 2);
-	if (!split)
-		return (NULL);
-	if (split[0])
-		split[0] = ft_strdup(dollar);
-	else
-	{
-		split[0] = NULL;
-		split[1] = NULL;
-	}
-	return (split);
-}
-
-char	**do_word_splitting(char *dollar, t_env *env)
-{
-	int		i;
-	char	*ifs;
-	char	**split;
-
-	i = 0;
-	split = NULL;
-	if (!get_env_value("IFS", env))
-		ifs = " \t\n";
-	else
-		ifs = "";
-	while (dollar[i])
-	{
-		if (is_ifs(dollar[i], env))
-			dollar[i] = *ifs;
-		i++;
-	}
-	if (ifs && strcmp(ifs, " \t\n") == 0)
-		split = ft_splitstr(dollar, ifs);
-	else if (ifs && strcmp(ifs, "") == 0)
-		split = do_ws_ifs(dollar, ifs);
-	if (!split)
-		split = case_no_ws(dollar);
-	else
-		free(dollar);
-	return (split);
-}
-
-void	cleanup_vars(char *before, char *after, char **splited, char **a_s)
-{
-	if (before)
-		free(before);
-	if (before)
-		free(before);
-	if (after)
-		free(after);
-	if (a_s)
-		free_matrix(a_s);
-	if (splited && splited[0] && splited[0][0])
-		free_matrix(splited);
-}
-
-char	*set_context_before(char **args, int i)
-{
-	char	*before;
-
-	if (args && args[i] && args[i][0] != '$')
-	{
-		before = set_before_dollar(args[i]);
-		if (before && before[0])
-			return (before);
-		if (before)
-			free (before);
-	}
-	return (NULL);
-}
 
 char	*set_context_after(char **args, int i)
 {
